Patient Creator Mutation
With the collaborative effort of me and kondanta we were able to fix the mutation for this feature. For the records, no thanks to flutter-graphql library and it's
wonderful documentation...

Also, UI is really bad but that's a placeholder so I am not going to fix it. At least for the near future.