implement: Doctor/Patient info for Patient/Doctor on RTC
the initial idea was implementing a page for patients which allows them to see their doctors and their information. I decided to put that on RTC. So whenever patient's doctor(s) available on RTC server, cards color turns to the green (they're available for video chat). If they're offline, cards color will be red and call button will be disabled.

I am also thinking that using this design as for doctor's dashboard. They'll see their patients information on these cards. There'll be an extra button that leads them to the patient's detailed information page.

Signed-off-by: Taylan Dogan <doganntaylan@gmail.com>